+++ approximation-properties-service/src/main/java/pl/leastsquaresalgorithms/approximationpropertiesservice/repository/ApproximationPropertiesRepository.java
@@ -4,13 +4,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import pl.leastsquaresalgorithms.approximationpropertiesservice.model.ApproximationPropertiesEntity;
 
-import java.math.BigInteger;
 import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface ApproximationPropertiesRepository extends JpaRepository<ApproximationPropertiesEntity, BigInteger> {
-    List<ApproximationPropertiesEntity> findByUserIdAndDeleted(BigInteger userId, Boolean deleted);
+public interface ApproximationPropertiesRepository extends JpaRepository<ApproximationPropertiesEntity, Long> {
+    List<ApproximationPropertiesEntity> findByUserIdAndDeleted(Long userId, Boolean deleted);
 
-    Optional<ApproximationPropertiesEntity> findByApproximationPropertiesIdAndDeleted(BigInteger approximationPropertiesId, Boolean deleted);
+    Optional<ApproximationPropertiesEntity> findByApproximationPropertiesIdAndDeleted(Long approximationPropertiesId, Boolean deleted);
 }
