+++ data-series-service/src/main/java/pl/leastsquaresalgorithms/dataseries/dto/DataSeriesFileDto.java
@@ -12,8 +12,8 @@ import java.util.List;
 @Setter
 @Builder
 public class DataSeriesFileDto {
-    private BigInteger id;
-    private BigInteger userId;
+    private Long id;
+    private Long userId;
     private String name;
     private String hashName;
     private Timestamp dateSent;
